major refactor, remove all unix-based stuff (should work on Windows, but haven't tested). This refactor is about issue 14 (Removing Unix-only features)

+++ src/constants.rs
@@ -1,6 +1,6 @@
 use home::home_dir;
 
-pub static DEFAULT_PORT: u32 = 17757;
+pub static DEFAULT_PORT: u16 = 17757;
 static DEFAULT_CONFIG_FILE_NAME: &str = ".horsetab.conf";
 
 pub static DEFAULT_INTERPRETER: &str = "/bin/sh";
